chore: convert auth.js to auth.ts
